Revert "improvement: Add most of missing JDK9+ methods for `java.nio` buffers (#3681)"
This reverts commit b07d16e2caa4c1fe62e9b73be980c3fff8f5809e.
