+++ javalib/src/main/scala/java/nio/HeapBuffers.scala
@@ -27,10 +27,6 @@ private[nio] final class HeapCharBuffer private (
     GenHeapBuffer[CharBuffer](this).generic_slice()
 
   @noinline
-  def slice(index: Int, length: Int): CharBuffer =
-    GenHeapBuffer[CharBuffer](this).generic_slice(index, length)
-
-  @noinline
   def duplicate(): CharBuffer =
     GenHeapBuffer[CharBuffer](this).generic_duplicate()
 
@@ -177,10 +173,6 @@ private[nio] final class HeapShortBuffer private (
     GenHeapBuffer[ShortBuffer](this).generic_slice()
 
   @noinline
-  def slice(index: Int, length: Int): ShortBuffer =
-    GenHeapBuffer[ShortBuffer](this).generic_slice(index, length)
-
-  @noinline
   def duplicate(): ShortBuffer =
     GenHeapBuffer[ShortBuffer](this).generic_duplicate()
 
@@ -313,10 +305,6 @@ private[nio] final class HeapIntBuffer private (
   @noinline
   def slice(): IntBuffer =
     GenHeapBuffer[IntBuffer](this).generic_slice()
-
-  @noinline
-  def slice(index: Int, length: Int): IntBuffer =
-    GenHeapBuffer[IntBuffer](this).generic_slice(index, length)
 
   @noinline
   def duplicate(): IntBuffer =
@@ -453,10 +441,6 @@ private[nio] final class HeapLongBuffer private (
     GenHeapBuffer[LongBuffer](this).generic_slice()
 
   @noinline
-  def slice(index: Int, length: Int): LongBuffer =
-    GenHeapBuffer[LongBuffer](this).generic_slice(index, length)
-
-  @noinline
   def duplicate(): LongBuffer =
     GenHeapBuffer[LongBuffer](this).generic_duplicate()
 
@@ -591,10 +575,6 @@ private[nio] final class HeapFloatBuffer private (
     GenHeapBuffer[FloatBuffer](this).generic_slice()
 
   @noinline
-  def slice(index: Int, length: Int): FloatBuffer =
-    GenHeapBuffer[FloatBuffer](this).generic_slice(index, length)
-
-  @noinline
   def duplicate(): FloatBuffer =
     GenHeapBuffer[FloatBuffer](this).generic_duplicate()
 
@@ -727,10 +707,6 @@ private[nio] final class HeapDoubleBuffer private (
   @noinline
   def slice(): DoubleBuffer =
     GenHeapBuffer[DoubleBuffer](this).generic_slice()
-
-  @noinline
-  def slice(index: Int, length: Int): DoubleBuffer =
-    GenHeapBuffer[DoubleBuffer](this).generic_slice(index, length)
 
   @noinline
   def duplicate(): DoubleBuffer =
