+++ javalib/src/main/scala/java/nio/HeapBuffers.scala.gyb
@@ -34,10 +34,6 @@ private[nio] final class Heap${T}Buffer private (
     GenHeapBuffer[${T}Buffer](this).generic_slice()
 
   @noinline
-  def slice(index: Int, length: Int): ${T}Buffer =
-    GenHeapBuffer[${T}Buffer](this).generic_slice(index, length)
-
-  @noinline
   def duplicate(): ${T}Buffer =
     GenHeapBuffer[${T}Buffer](this).generic_duplicate()
