+++ javalib/src/main/scala/java/nio/HeapByteBuffer.scala
@@ -31,10 +31,6 @@ private[nio] class HeapByteBuffer(
     genHeapBuffer.generic_slice()
 
   @noinline
-  def slice(index: Int, length: Int): ByteBuffer =
-    genHeapBuffer.generic_slice(index, length)
-
-  @noinline
   def duplicate(): ByteBuffer =
     genHeapBuffer.generic_duplicate()
