+++ javalib/src/main/scala/java/nio/HeapByteBufferViews.scala.gyb
@@ -35,10 +35,6 @@ private[nio] final class HeapByteBuffer${T}View private (
     GenHeapBufferView[${T}Buffer](this).generic_slice()
 
   @noinline
-  def slice(index: Int, length: Int): ${T}Buffer =
-    GenHeapBufferView[${T}Buffer](this).generic_slice(index, length)
-
-  @noinline
   def duplicate(): ${T}Buffer =
     GenHeapBufferView[${T}Buffer](this).generic_duplicate()
