+++ javalib/src/main/scala/java/nio/MappedByteBufferImpl.scala
@@ -70,10 +70,6 @@ private class MappedByteBufferImpl(
     genMappedBuffer.generic_slice()
 
   @noinline
-  def slice(index: Int, length: Int): ByteBuffer =
-    genMappedBuffer.generic_slice(index, length)
-
-  @noinline
   def duplicate(): ByteBuffer =
     genMappedBuffer.generic_duplicate()
