+++ javalib/src/main/scala/java/nio/MappedByteBufferViews.scala
@@ -28,10 +28,6 @@ private[nio] final class MappedByteBufferCharView private (
     GenMappedBufferView[CharBuffer](this).generic_slice()
 
   @noinline
-  def slice(index: Int, length: Int): CharBuffer =
-    GenMappedBufferView[CharBuffer](this).generic_slice(index, length)
-
-  @noinline
   def duplicate(): CharBuffer =
     GenMappedBufferView[CharBuffer](this).generic_duplicate()
 
@@ -154,10 +150,6 @@ private[nio] final class MappedByteBufferShortView private (
     GenMappedBufferView[ShortBuffer](this).generic_slice()
 
   @noinline
-  def slice(index: Int, length: Int): ShortBuffer =
-    GenMappedBufferView[ShortBuffer](this).generic_slice(index, length)
-
-  @noinline
   def duplicate(): ShortBuffer =
     GenMappedBufferView[ShortBuffer](this).generic_duplicate()
 
@@ -265,10 +257,6 @@ private[nio] final class MappedByteBufferIntView private (
   @noinline
   def slice(): IntBuffer =
     GenMappedBufferView[IntBuffer](this).generic_slice()
-
-  @noinline
-  def slice(index: Int, length: Int): IntBuffer =
-    GenMappedBufferView[IntBuffer](this).generic_slice(index, length)
 
   @noinline
   def duplicate(): IntBuffer =
@@ -380,10 +368,6 @@ private[nio] final class MappedByteBufferLongView private (
     GenMappedBufferView[LongBuffer](this).generic_slice()
 
   @noinline
-  def slice(index: Int, length: Int): LongBuffer =
-    GenMappedBufferView[LongBuffer](this).generic_slice(index, length)
-
-  @noinline
   def duplicate(): LongBuffer =
     GenMappedBufferView[LongBuffer](this).generic_duplicate()
 
@@ -493,10 +477,6 @@ private[nio] final class MappedByteBufferFloatView private (
     GenMappedBufferView[FloatBuffer](this).generic_slice()
 
   @noinline
-  def slice(index: Int, length: Int): FloatBuffer =
-    GenMappedBufferView[FloatBuffer](this).generic_slice(index, length)
-
-  @noinline
   def duplicate(): FloatBuffer =
     GenMappedBufferView[FloatBuffer](this).generic_duplicate()
 
@@ -604,10 +584,6 @@ private[nio] final class MappedByteBufferDoubleView private (
   @noinline
   def slice(): DoubleBuffer =
     GenMappedBufferView[DoubleBuffer](this).generic_slice()
-
-  @noinline
-  def slice(index: Int, length: Int): DoubleBuffer =
-    GenMappedBufferView[DoubleBuffer](this).generic_slice(index, length)
 
   @noinline
   def duplicate(): DoubleBuffer =
