+++ javalib/src/main/scala/java/nio/PointerByteBufferViews.scala
@@ -30,10 +30,6 @@ private[nio] final class PointerByteBufferCharView private (
     GenPointerBufferView[CharBuffer](this).generic_slice()
 
   @noinline
-  def slice(index: Int, length: Int): CharBuffer =
-    GenPointerBufferView[CharBuffer](this).generic_slice(index, length)
-
-  @noinline
   def duplicate(): CharBuffer =
     GenPointerBufferView[CharBuffer](this).generic_duplicate()
 
@@ -156,10 +152,6 @@ private[nio] final class PointerByteBufferShortView private (
     GenPointerBufferView[ShortBuffer](this).generic_slice()
 
   @noinline
-  def slice(index: Int, length: Int): ShortBuffer =
-    GenPointerBufferView[ShortBuffer](this).generic_slice(index, length)
-
-  @noinline
   def duplicate(): ShortBuffer =
     GenPointerBufferView[ShortBuffer](this).generic_duplicate()
 
@@ -267,10 +259,6 @@ private[nio] final class PointerByteBufferIntView private (
   @noinline
   def slice(): IntBuffer =
     GenPointerBufferView[IntBuffer](this).generic_slice()
-
-  @noinline
-  def slice(index: Int, length: Int): IntBuffer =
-    GenPointerBufferView[IntBuffer](this).generic_slice(index, length)
 
   @noinline
   def duplicate(): IntBuffer =
@@ -382,10 +370,6 @@ private[nio] final class PointerByteBufferLongView private (
     GenPointerBufferView[LongBuffer](this).generic_slice()
 
   @noinline
-  def slice(index: Int, length: Int): LongBuffer =
-    GenPointerBufferView[LongBuffer](this).generic_slice(index, length)
-
-  @noinline
   def duplicate(): LongBuffer =
     GenPointerBufferView[LongBuffer](this).generic_duplicate()
 
@@ -495,10 +479,6 @@ private[nio] final class PointerByteBufferFloatView private (
     GenPointerBufferView[FloatBuffer](this).generic_slice()
 
   @noinline
-  def slice(index: Int, length: Int): FloatBuffer =
-    GenPointerBufferView[FloatBuffer](this).generic_slice(index, length)
-
-  @noinline
   def duplicate(): FloatBuffer =
     GenPointerBufferView[FloatBuffer](this).generic_duplicate()
 
@@ -606,10 +586,6 @@ private[nio] final class PointerByteBufferDoubleView private (
   @noinline
   def slice(): DoubleBuffer =
     GenPointerBufferView[DoubleBuffer](this).generic_slice()
-
-  @noinline
-  def slice(index: Int, length: Int): DoubleBuffer =
-    GenPointerBufferView[DoubleBuffer](this).generic_slice(index, length)
 
   @noinline
   def duplicate(): DoubleBuffer =
