+++ javalib/src/main/scala/java/nio/PointerByteBufferViews.scala.gyb
@@ -37,10 +37,6 @@ private[nio] final class PointerByteBuffer${T}View private (
     GenPointerBufferView[${T}Buffer](this).generic_slice()
 
   @noinline
-  def slice(index: Int, length: Int): ${T}Buffer =
-    GenPointerBufferView[${T}Buffer](this).generic_slice(index, length)
-
-  @noinline
   def duplicate(): ${T}Buffer =
     GenPointerBufferView[${T}Buffer](this).generic_duplicate()
