+++ scripts/gyb_all.sh
@@ -39,10 +39,9 @@ gyb_files $nativelib $scala $runtime Arrays Boxes Primitives
 gyb clib/src/main/scala/scala/scalanative/libc/stdatomic.scala.gyb
 gyb clib/src/main/resources/scala-native/stdatomic.c.gyb
 
-gyb $javaNIO/Buffers.scala.gyb
 gyb $javaNIO/HeapBuffers.scala.gyb
 gyb $javaNIO/HeapByteBufferViews.scala.gyb
 gyb $javaNIO/MappedByteBufferViews.scala.gyb
 gyb $javaNIO/PointerByteBufferViews.scala.gyb
 
-gyb unit-tests/shared/src/test/require-jdk16/org/scalanative/testsuite/javalib/nio/ByteBufferViewsNativeTests.scala.gyb
+gyb unit-tests/native/src/test/scala/org/scalanative/testsuite/niobuffer/ByteBufferViewsNativeTests.scala.gyb
