+++ unit-tests/jvm/src/test/scala/org/scalanative/testsuite/javalib/nio/BaseBufferPlatformTest.scala.scala
@@ -1,0 +1,3 @@
+package org.scalanative.testsuite.javalib.nio
+
+trait BaseBufferPlatformTest { self: BaseBufferTest => }
