+++ unit-tests/native/src/test/scala/org/scalanative/testsuite/javalib/nio/BaseBufferPlatformTest.scala.scala
@@ -2,26 +2,26 @@ package org.scalanative.testsuite.javalib.nio
 
 import java.nio._
 
+// Ported from Scala.js
+
 import org.junit.Test
 import org.junit.Assert._
-import org.scalanative.testsuite.utils.Platform._
 
-// Extended Scala Native API
 import scala.scalanative.memory.PointerBuffer
 import scala.scalanative.memory.PointerBufferOps._
 
-trait ScalaNativeBufferTests { self: BaseBufferTest =>
-  // TODO: It should not be shared with JVM, but it's required due to BaseBuffer coupling with JDK8
-
+trait BaseBufferPlatformTest { self: BaseBufferTest =>
   import factory._
-  @Test def hasPointer(): Unit = if (executingInScalaNative) {
+
+  // Extended Scala Native API
+  @Test def hasPointer(): Unit = {
     val buf = factory.allocBuffer(8)
     if (createsReadOnly)
       assertFalse("read-only, access to pointer", buf.hasPointer())
     else assertEquals("hasPointer", createsPointerBuffer, buf.hasPointer())
   }
 
-  @Test def getPointer(): Unit = if (executingInScalaNative) {
+  @Test def getPointer(): Unit = {
     val buf = factory.allocBuffer(8)
     if (createsReadOnly || !createsPointerBuffer)
       assertThrows(
@@ -30,5 +30,4 @@ trait ScalaNativeBufferTests { self: BaseBufferTest =>
       )
     else assertNotNull(buf.pointer())
   }
-
 }
