+++ unit-tests/shared/src/test/scala/org/scalanative/testsuite/javalib/nio/BaseBufferTest.scala
@@ -11,8 +11,7 @@ import org.junit.Assume._
 import org.scalanative.testsuite.utils.AssertThrows.assertThrows
 import org.scalanative.testsuite.utils.Platform._
 
-// Most of this tests should not be require JDK 16, but decoupling the BaseBuffer tests infrastructure is way to hard right now
-abstract class BaseBufferTest extends ScalaNativeBufferTests {
+abstract class BaseBufferTest extends BaseBufferPlatformTest {
 
   type Factory <: BufferFactory
 
@@ -400,4 +399,5 @@ abstract class BaseBufferTest extends ScalaNativeBufferTests {
       assertEquals(elemFromInt(23), buf2.get(6))
     }
   }
+
 }
