+++ unit-tests/shared/src/test/scala/org/scalanative/testsuite/javalib/nio/FloatBufferTest.scala
